+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# (Wed Mar 29 2023)
+
+#### 🚀 Enhancement
+
+- `@magic-sdk/types@13.1.0`
+  - Adds `MagicBoxHeartBeat` to MagicPayloadMethod [#486](https://github.com/magiclabs/magic-js/pull/486) ([@Ariflo](https://github.com/Ariflo))
+
+#### Authors: 1
+
+- Arian Flores ([@Ariflo](https://github.com/Ariflo))
+
+---
+
+# (Mon Mar 27 2023)
+
+#### 💥 Breaking Change
+
+- `@magic-sdk/provider@15.0.0`, `@magic-sdk/types@13.0.0`
+  - Revert "move third party wallet connections to sdk (#473)" [#480](https://github.com/magiclabs/magic-js/pull/480) ([@Ariflo](https://github.com/Ariflo))
+
+#### Authors: 1
+
+- Arian Flores ([@Ariflo](https://github.com/Ariflo))
+
+---
+
 # (Fri Mar 24 2023)
 
 #### 🐛 Bug Fix